Find additional errors
